Rename `CallbackManager` to `CallbackDispatcher`.
