+++ java/client/src/main/java/babushka/connectors/handlers/ChannelHandler.java
@@ -1,7 +1,6 @@
 package babushka.connectors.handlers;
 
 import babushka.connectors.resources.Platform;
-import babushka.managers.CallbackManager;
 import connection_request.ConnectionRequestOuterClass.ConnectionRequest;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.Channel;
@@ -14,29 +13,29 @@ import response.ResponseOuterClass.Response;
 
 /**
  * Class responsible for manipulations with Netty's {@link Channel}.<br>
- * Uses a {@link CallbackManager} to record callbacks of every request sent.
+ * Uses a {@link CallbackDispatcher} to record callbacks of every request sent.
  */
 public class ChannelHandler {
   private final Channel channel;
-  private final CallbackManager callbackManager;
+  private final CallbackDispatcher callbackDispatcher;
 
   /** Open a new channel for a new client. */
-  public ChannelHandler(CallbackManager callbackManager, String socketPath) {
+  public ChannelHandler(CallbackDispatcher callbackDispatcher, String socketPath) {
     channel =
         new Bootstrap()
             // TODO let user specify the thread pool or pool size as an option
             .group(Platform.createNettyThreadPool("babushka-channel", OptionalInt.empty()))
             .channel(Platform.getClientUdsNettyChannelType())
-            .handler(new ProtobufSocketChannelInitializer(callbackManager))
+            .handler(new ProtobufSocketChannelInitializer(callbackDispatcher))
             .connect(new DomainSocketAddress(socketPath))
             // TODO call here .sync() if needed or remove this comment
             .channel();
-    this.callbackManager = callbackManager;
+    this.callbackDispatcher = callbackDispatcher;
   }
 
   /** Write a protobuf message to the socket. */
   public CompletableFuture<Response> write(RedisRequest.Builder request, boolean flush) {
-    var commandId = callbackManager.registerRequest();
+    var commandId = callbackDispatcher.registerRequest();
     request.setCallbackIdx(commandId.getKey());
 
     if (flush) {
@@ -50,7 +49,7 @@ public class ChannelHandler {
   /** Write a protobuf message to the socket. */
   public CompletableFuture<Response> connect(ConnectionRequest request) {
     channel.writeAndFlush(request);
-    return callbackManager.getConnectionPromise();
+    return callbackDispatcher.registerConnection();
   }
 
   private final AtomicBoolean closed = new AtomicBoolean(false);
@@ -59,7 +58,7 @@ public class ChannelHandler {
   public void close() {
     if (closed.compareAndSet(false, true)) {
       channel.close();
-      callbackManager.shutdownGracefully();
+      callbackDispatcher.shutdownGracefully();
     }
   }
 }
